refactor(daemon): split daemon manager by lifecycle area
